feat: update candle client
